requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
datasets==2.17.1
dill==0.3.8
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.21.3
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4

langchain~=0.1.11
python-dotenv~=1.0.1